pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.14",
    "jose>=1.0.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
]
